+++ CHANGELOG.md
@@ -12,7 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - added example check_file plugin - will be improved to send files to ClamAV
 - added option to use ssh private keys for remote authentication
 - added option to provide remote ssh server fingerprints
-- added new authentication plugin "AuthenticatorRemote" which allows to provide remote credentials without passthrough 
+- added new authentication plugin "AuthenticatorRemote" which allows to provide remote credentials without passthrough
 
 ### Fixed
 
@@ -28,6 +28,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - update paramiko requirements to 4.0
 - removed support for DSS keys
 - replaced sshpubkeys module with paramiko based utility class to handle ssh public keys
+- updated documentation for the next relese of SSH-MITM
 
 
 ## [5.0.1] - 2025-01-22